requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
mysqlclient==2.2.6
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
